Move rumEventsCollection/resource
