+++ packages/rum-core/src/boot/startRum.ts
@@ -19,7 +19,7 @@ import { startRequestCollection } from '../domain/requestCollection'
 import { startActionCollection } from '../domain/action/actionCollection'
 import { startErrorCollection } from '../domain/error/errorCollection'
 import { startLongTaskCollection } from '../domain/longTask/longTaskCollection'
-import { startResourceCollection } from '../domain/rumEventsCollection/resource/resourceCollection'
+import { startResourceCollection } from '../domain/resource/resourceCollection'
 import { startViewCollection } from '../domain/rumEventsCollection/view/viewCollection'
 import type { RumSessionManager } from '../domain/rumSessionManager'
 import { startRumSessionManager, startRumSessionManagerStub } from '../domain/rumSessionManager'
