+++ packages/rum-core/src/domain/resource/matchRequestTiming.spec.ts
@@ -1,8 +1,8 @@
 import type { Duration, RelativeTime } from '@datadog/browser-core'
 import { isIE, relativeToClocks } from '@datadog/browser-core'
-import { createResourceEntry } from '../../../../test'
-import type { RumPerformanceResourceTiming } from '../../../browser/performanceCollection'
-import type { RequestCompleteEvent } from '../../requestCollection'
+import { createResourceEntry } from '../../../test'
+import type { RumPerformanceResourceTiming } from '../../browser/performanceCollection'
+import type { RequestCompleteEvent } from '../requestCollection'
 
 import { matchRequestTiming } from './matchRequestTiming'
