+++ packages/rum-core/src/domain/resource/matchRequestTiming.ts
@@ -1,7 +1,7 @@
 import type { Duration, RelativeTime } from '@datadog/browser-core'
 import { addDuration } from '@datadog/browser-core'
-import type { RumPerformanceResourceTiming } from '../../../browser/performanceCollection'
-import type { RequestCompleteEvent } from '../../requestCollection'
+import type { RumPerformanceResourceTiming } from '../../browser/performanceCollection'
+import type { RequestCompleteEvent } from '../requestCollection'
 import { toValidEntry } from './resourceUtils'
 
 interface Timing {
