+++ packages/rum-core/src/domain/resource/resourceCollection.ts
@@ -11,21 +11,21 @@ import {
   ExperimentalFeature,
 } from '@datadog/browser-core'
 import type { ClocksState, Duration } from '@datadog/browser-core'
-import type { RumConfiguration } from '../../configuration'
-import type { RumPerformanceEntry, RumPerformanceResourceTiming } from '../../../browser/performanceCollection'
+import type { RumConfiguration } from '../configuration'
+import type { RumPerformanceEntry, RumPerformanceResourceTiming } from '../../browser/performanceCollection'
 import type {
   PerformanceEntryRepresentation,
   RumXhrResourceEventDomainContext,
   RumFetchResourceEventDomainContext,
-} from '../../../domainContext.types'
-import type { RawRumResourceEvent } from '../../../rawRumEvent.types'
-import { RumEventType } from '../../../rawRumEvent.types'
-import type { LifeCycle, RawRumEventCollectedData } from '../../lifeCycle'
-import { LifeCycleEventType } from '../../lifeCycle'
-import type { RequestCompleteEvent } from '../../requestCollection'
-import type { RumSessionManager } from '../../rumSessionManager'
-import type { PageStateHistory } from '../../contexts/pageStateHistory'
-import { PageState } from '../../contexts/pageStateHistory'
+} from '../../domainContext.types'
+import type { RawRumResourceEvent } from '../../rawRumEvent.types'
+import { RumEventType } from '../../rawRumEvent.types'
+import type { LifeCycle, RawRumEventCollectedData } from '../lifeCycle'
+import { LifeCycleEventType } from '../lifeCycle'
+import type { RequestCompleteEvent } from '../requestCollection'
+import type { RumSessionManager } from '../rumSessionManager'
+import type { PageStateHistory } from '../contexts/pageStateHistory'
+import { PageState } from '../contexts/pageStateHistory'
 import { matchRequestTiming } from './matchRequestTiming'
 import {
   computePerformanceResourceDetails,
