+++ packages/rum-core/src/domain/resource/resourceUtils.ts
@@ -10,10 +10,10 @@ import {
   toServerDuration,
 } from '@datadog/browser-core'
 
-import type { RumPerformanceResourceTiming } from '../../../browser/performanceCollection'
+import type { RumPerformanceResourceTiming } from '../../browser/performanceCollection'
 
-import type { PerformanceResourceDetailsElement } from '../../../rawRumEvent.types'
-import type { RumConfiguration } from '../../configuration'
+import type { PerformanceResourceDetailsElement } from '../../rawRumEvent.types'
+import type { RumConfiguration } from '../configuration'
 
 export interface PerformanceResourceDetails {
   redirect?: PerformanceResourceDetailsElement
